SmartNoteProcessor/requirements.txt
email-validator>=2.2.0
Flask
Flask-SQLAlchemy>=3.1.1
google-generativeai>=0.8.4
gunicorn>=23.0.0
pdfminer-six>=20250327
Pillow
psycopg2-binary>=2.9.10
PyMuPDF>=1.25.5
pypdf2>=3.0.1
python-docx>=1.1.2
python-dotenv>=0.19
SQLAlchemy>=2.0.40
Werkzeug>=3.1.3
moviepy==1.0.3
openai-whisper